Convert dsMutation to TypeScript
