+++ metaspace/graphql/executableSchema.ts
@@ -9,6 +9,7 @@ import {Resolvers as UserResolvers} from './src/modules/user/controller';
 import {Resolvers as GroupResolvers} from './src/modules/group/controller';
 import {Resolvers as SystemResolvers} from './src/modules/system/controller';
 import {Resolvers as ProjectResolvers} from './src/modules/project/controller';
+import {Resolvers as DatasetResolvers} from './src/modules/dataset/controller';
 import * as Resolvers from './resolvers';
 import {mergedSchemas} from './schema';
 import addReadOnlyInterceptorToSchema from './src/modules/system/addReadOnlyInterceptorToSchema';
@@ -24,6 +25,7 @@ export const makeNewExecutableSchema = () => {
       GroupResolvers,
       ProjectResolvers,
       SystemResolvers,
+      DatasetResolvers,
     ],
   });
 };
