+++ metaspace/graphql/src/modules/dataset/controller/index.ts
@@ -1,0 +1,7 @@
+import Mutation from './Mutation';
+import {IResolvers} from 'graphql-tools';
+import {Context} from '../../../context';
+
+export const Resolvers = {
+  Mutation
+} as IResolvers<any, Context>;
